requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt setup.py
#
ariadne==0.13.0
    # via ariadne-relay (setup.py)
graphql-core==3.1.5
    # via
    #   ariadne
    #   graphql-relay
graphql-relay==3.1.0
    # via ariadne-relay (setup.py)
starlette==0.14.2
    # via ariadne
typing-extensions==3.10.0.0
    # via ariadne
